requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
twilio==8.10.0
groq==0.4.1
cartesia==1.0.3
deepgram-sdk==3.2.6
python-dotenv==1.0.0
pydantic==2.5.0
asyncio-mqtt==0.16.1
aiofiles==23.2.1
python-multipart==0.0.6
httpx==0.25.2